+++ releasenotes.md
@@ -2,6 +2,7 @@
 
 ## Version 0.16.0
 ### New Features
+- IDNumber.Invalid()
 - IDNumber.Valid()
 
 ## Version 0.15.0